transmit oopdated!!!!

+++ src/protocol.h
@@ -10,13 +10,14 @@ const char ACK = 0x06;
 const char RVI = 0x17;
 const char NAK = 0x21;
 
-const int PACKET_SIZE = 1024; //packet size in bytes
+const unsigned PACKET_SIZE = 1024; //packet size in bytes
 
 struct WConn
 {
 	HANDLE hComm;
 
-	double		TO1,
+	unsigned
+		TO1,
 		TO2,
 		TO3,
 		TO4
